LSG_BACKUP_2026-03-11/scripts/requirements.txt
# Little Smart Genius — Python Dependencies
# Install: pip install -r requirements.txt

requests>=2.28.0
Pillow>=9.0.0
python-dotenv>=0.19.0
aiohttp>=3.9.0
